[Feat] level 3-1 enum, 제네릭, 큰값 조회(lamda)  #24

+++ level3/nbcamp-java-personal-task/src/main/java/calculator/ModOperator.java
@@ -1,7 +1,7 @@
 package calculator;
 
 public class ModOperator implements Operator{
-    public double operate(int num1, int num2) {
+    public double operate(double num1, double num2) {
         return (double) num1 % num2;
     }
 }
